storage: move raft processing into own file
Release note: None
